+++ feature/booking/common/presentation/src/main/java/com/softserveacademy/feature/booking/common/presentation/ui/screens/TravelBookingConfirmScreen.kt
@@ -23,7 +23,7 @@ import com.softserveacademy.core.presentation.design_system.components.TravelErr
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun TravelBookingConfirmScreen(
-    totalPrice: Int,
+    totalPrice: Double,
     onBackClick: () -> Unit,
     onConfirmClick: () -> Unit,
     modifier: Modifier = Modifier,
@@ -78,7 +78,7 @@ fun TravelBookingConfirmScreen(
 
 @Composable
 fun TravelBookingConfirmBottomBar(
-    totalPrice: Int,
+    totalPrice: Double,
     onButtonClick: () -> Unit,
     modifier: Modifier = Modifier,
     isLoading: Boolean = false
@@ -136,7 +136,7 @@ fun TravelBookingConfirmBottomBar(
 fun TravelBookingConfirmScreenPreview() {
     Travelin2026ProjectLabTheme() {
         TravelBookingConfirmScreen(
-            totalPrice = 100,
+            totalPrice = 100.0,
             onBackClick = {},
             onConfirmClick = {},
             content = {
