+++ feature/booking/tour/presentation/src/main/java/com/softserveacademy/feature/booking/tour/presentation/ui/screens/TourBookingConfirmScreen.kt
@@ -176,7 +176,7 @@ fun TourBookingConfirmContent(
         TravelLoadingScreen()
     } else {
         TravelBookingConfirmScreen(
-            totalPrice = uiState.totalPrice.toInt(),
+            totalPrice = uiState.totalPrice,
             onBackClick = onBackClick,
             onConfirmClick = onConfirmClick,
             isConfirmLoading = uiState.isPaymentSheetLoading,
